DDF-1748: Combine all Configuration related integration tests into a single class
